blip2-project/requirements.txt
Flask
salesforce-lavis
torch
numpy
opencv-python
pillow
python-dotenv